+++ docs/CHANGELOG.rst
@@ -6,8 +6,9 @@ Changelog
 0.9.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Infocard have no more a location field but a servicetype one (#53)
+- Clean up MANIFEST.in.
+[ale-rt]
 
 0.9.1 (2014-11-06)
 ------------------